backend/requirements.txt
mysql-connector-python
flask
flask-cors
bcrypt
PyJWT
torch
transformers
PyMuPDF
sentencepiece
requests
python-dotenv
